# openapi.json
{"openapi":"3.1.0","info":{"title":"Legal Metrology Compliance Verification API","description":"Automated software system to assess compliance of packaged commodities under the Legal Metrology (Packaged Commodities) Rules, 2011.","version":"1.0.0"},"paths":{"/":{"get":{"tags":["General"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["General"],"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/scans/analyze":{"post":{"tags":["Scans & Compliance Verification"],"summary":"Analyze Package","description":"Receives package label image and dimensions, runs OCR & compliance engine,\nand returns parsed fields along with identified statutory violations.","operationId":"analyze_package_api_v1_scans_analyze_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_analyze_package_api_v1_scans_analyze_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/scans/history":{"get":{"tags":["Scans & Compliance Verification"],"summary":"Get Scan History","description":"Returns list of all scanned packaging records for the audit log.","operationId":"get_scan_history_api_v1_scans_history_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/scans/{scan_id}":{"get":{"tags":["Scans & Compliance Verification"],"summary":"Get Scan By Id","description":"Retrieves a specific scan result by its unique UUID.","operationId":"get_scan_by_id_api_v1_scans__scan_id__get","parameters":[{"name":"scan_id","in":"path","required":true,"schema":{"type":"string","title":"Scan Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Authentication & Users"],"summary":"Login","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_api_v1_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/rules/summary":{"get":{"tags":["Legal Metrology Rules"],"summary":"Get All Rules Summary","description":"Returns dataset metadata, active amendments, and all 34 primary statutory rules.","operationId":"get_all_rules_summary_api_v1_rules_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/rules/requirements":{"get":{"tags":["Legal Metrology Rules"],"summary":"Get Mandatory Requirements","description":"Fetches mandatory Rule 6 field declarations, permitted regex patterns, and units.","operationId":"get_mandatory_requirements_api_v1_rules_requirements_get","parameters":[{"name":"severity","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by severity: HIGH, MEDIUM, LOW","title":"Severity"},"description":"Filter by severity: HIGH, MEDIUM, LOW"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/rules/font-size-matrix":{"get":{"tags":["Legal Metrology Rules"],"summary":"Get Font Size Matrix","description":"Fetches Rule 7 Principal Display Panel area vs font height threshold lookup table.","operationId":"get_font_size_matrix_api_v1_rules_font_size_matrix_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/rules/calculate-min-font":{"get":{"tags":["Legal Metrology Rules"],"summary":"Calculate Required Font","description":"Calculates the statutory minimum font height required for specific package dimensions.","operationId":"calculate_required_font_api_v1_rules_calculate_min_font_get","parameters":[{"name":"pdp_area_cm2","in":"query","required":true,"schema":{"type":"number","description":"Surface area of PDP in cm²","title":"Pdp Area Cm2"},"description":"Surface area of PDP in cm²"},{"name":"net_quantity","in":"query","required":true,"schema":{"type":"number","description":"Net declared numeric quantity","title":"Net Quantity"},"description":"Net declared numeric quantity"},{"name":"unit","in":"query","required":true,"schema":{"type":"string","description":"Unit of measurement (g, kg, ml, l, etc.)","title":"Unit"},"description":"Unit of measurement (g, kg, ml, l, etc.)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/rules/exemptions":{"get":{"tags":["Legal Metrology Rules"],"summary":"Get Statutory Exemptions","description":"Fetches statutory exemptions under Rule 26 (small packs, bulk packs, DPCO drugs, etc.).","operationId":"get_statutory_exemptions_api_v1_rules_exemptions_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/rules/mpe-table":{"get":{"tags":["Legal Metrology Rules"],"summary":"Get Maximum Permissible Error Table","description":"Fetches the First Schedule Maximum Permissible Error (MPE) weight/volume thresholds.","operationId":"get_maximum_permissible_error_table_api_v1_rules_mpe_table_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/reports/show-cause-notice/{scan_id}":{"get":{"tags":["Inspection Reports & Notices"],"summary":"Generate Show Cause Notice","description":"Generates a statutory Show-Cause Legal Notice under Rule 32 of Legal Metrology Rules, 2011.","operationId":"generate_show_cause_notice_api_v1_reports_show_cause_notice__scan_id__get","parameters":[{"name":"scan_id","in":"path","required":true,"schema":{"type":"string","title":"Scan Id"}},{"name":"format","in":"query","required":false,"schema":{"type":"string","pattern":"^(json|html)$","default":"json","title":"Format"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reports/summary/{scan_id}":{"get":{"tags":["Inspection Reports & Notices"],"summary":"Get Scan Report Json","description":"Provides complete JSON audit report suitable for downstream PDF generation.","operationId":"get_scan_report_json_api_v1_reports_summary__scan_id__get","parameters":[{"name":"scan_id","in":"path","required":true,"schema":{"type":"string","title":"Scan Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analytics/dashboard-stats":{"get":{"tags":["Enforcement Dashboard Analytics"],"summary":"Get Dashboard Analytics","description":"Computes real-time compliance metrics, violation frequency distribution,\nand average compliance scores for the officer enforcement dashboard.","operationId":"get_dashboard_analytics_api_v1_analytics_dashboard_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/analytics/top-violations":{"get":{"tags":["Enforcement Dashboard Analytics"],"summary":"Get Top Violations","description":"Returns the most frequently breached Legal Metrology clauses across all scans.","operationId":"get_top_violations_api_v1_analytics_top_violations_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_analyze_package_api_v1_scans_analyze_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"height_cm":{"type":"number","title":"Height Cm","default":10.0},"width_cm":{"type":"number","title":"Width Cm","default":10.0},"pdp_type":{"type":"string","title":"Pdp Type","default":"rectangular"},"brand_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Brand Name"},"commodity_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Commodity Name"},"inspector_id":{"type":"string","title":"Inspector Id","default":"INSP-DEFAULT"}},"type":"object","required":["file"],"title":"Body_analyze_package_api_v1_scans_analyze_post"},"Body_login_api_v1_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_api_v1_auth_login_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"},"role":{"type":"string","title":"Role"},"username":{"type":"string","title":"Username"}},"type":"object","required":["access_token","token_type","role","username"],"title":"Token"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}